Added commands to originate fibbing routes
Fibbing routes are Type-5 LSA whose forwarding address is never
0.0.0.0, and can denote any host in the network.
The main command to add such route is located in the
OSPF config node:
fibbing A.B.C.D/M via A.B.C.D {cost <0-a lot> | metric-type (1|2)}
This instruction creates a route/LSA for the given prefix,
with the given IPv4 address as forwarding address.
Optionally, one can specify a custom cost and/or metric-type
to be used in the LSA, if he does not want to use the default
values specified via
redistribute fibbing {metric <0-a lot> | metric-type (1|2)}
Without the above statement, fibbing routes will not be redistributed.

For convenience, the fibbing command is also aliased to:
ip ospf fibbing _params_
Which will originate the fibbing route on the first OSPF instance,
or create one if none exists yet.

Signed-off-by: Olivier Tilmans <otilmans@gmail.com>